Rename spack.util.downloader -> spack.util.download
