+++ lib/spack/spack/util/download.py
